style: 커밋 컨벤션을 intellij-java-wooteco-style.xml 로 변경 후 적용
