+++ src/main/java/racingCar/Application.java
@@ -3,7 +3,7 @@ package racingCar;
 import racingCar.controller.GameController;
 
 public class Application {
-	public static void main(String[] args) {
-		new GameController().run();
-	}
+    public static void main(String[] args) {
+        new GameController().run();
+    }
 }
