+++ src/main/java/racingCar/model/Car.java
@@ -1,23 +1,23 @@
 package racingCar.model;
 
-import racingCar.utlis.Util;
+import racingCar.model.utlis.Util;
 
 public class Car {
-	public final Name name;
-	public final Position position = new Position();
+    public final Name name;
+    public final Position position = new Position();
 
-	public Car(String name) {
-		this.name = new Name(name);
-	}
+    public Car(String name) {
+        this.name = new Name(name);
+    }
 
-	public void go() {
-		final int goNumber = 4;
-		if (Util.getRandomInteger() >= goNumber) {
-			move();
-		}
-	}
+    public void go() {
+        final int goNumber = 4;
+        if (Util.getRandomInteger() >= goNumber) {
+            move();
+        }
+    }
 
-	public void move() {
-		position.move();
-	}
+    public void move() {
+        position.move();
+    }
 }
