+++ src/main/java/racingCar/model/Position.java
@@ -1,17 +1,17 @@
 package racingCar.model;
 
 public class Position {
-	private int position = 0;
+    private int position = 0;
 
-	public void move() {
-		position++;
-	}
+    public void move() {
+        position++;
+    }
 
-	public int get() {
-		return position;
-	}
+    public int get() {
+        return position;
+    }
 
-	public boolean isSame(int position) {
-		return this.position == position;
-	}
+    public boolean isSame(int position) {
+        return this.position == position;
+    }
 }
